Thala for a reason/requirements.txt
streamlit
itertools
pillow